python_tools/pyproject.toml
[project]
name = "python-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rich>=14.3.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
